inventory rewrite

+++ src/server/entity/tradepeer.cpp
@@ -148,10 +148,10 @@ void TradePeer::debugPrint()
   printf("====================================================\n");
   printf("TradePeer of %s\n", *getEntity()->getCharacter()->getName());
   printf("----------------------------------------------------\n");
-  printf("Session: %x\n", session);
+  printf("Session: %x\n", (intptr_t) session);
   if (session)
   {
-    printf("Session peers: 1) %x \t 2) %x\n", session->peer1, session->peer2);
+    printf("Session peers: 1) %x \t 2) %x\n", (intptr_t) session->peer1, (intptr_t) session->peer2);
   }
   printf("----------------------------------------------------\n");
 }
